requirements.txt
discord_webhook==0.12.0
pywin32==301
requests==2.26.0
qrcode==6.1
Pillow==8.4.0
pycryptodome==3.12.0
websocket_client==1.2.3
